Refactor getters into own class
